Added SQL query to add extra database column
Added in a SQL query to add in a database column upon updating the
shoutbox, if the version is 1.1.3 or below.

Removed the "updates" folder with all the .sql files in.

Removed the update tags in the XML file as using SQL files to update the
database would not have worked as there are no SQL files in the previous
version.